tree-construction/tests6.dat

#data
<!doctype html></head> <head>
#errors
Unexpected start tag head. Ignored.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   " "
|   <body>

#data
<!doctype html></html> <head>
#errors
Unexpected start tag head.
Unexpected start tag head in after body phase.
Unexpected start tag head. Ignored.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " "

#data
<!doctype html></body><meta>
#errors
Unexpected meta element in after body phase.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <meta>

#data
<!doctype HTml><form><div></form><div>
#errors
Form end tag ignored.
Unexpected end of file.
#document
| <!DOCTYPE HTml>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         <div>

#data
<!doctype HTml><title>&amp;</title>
#errors
#document
| <!DOCTYPE HTml>
| <html>
|   <head>
|     <title>
|       "&"
|   <body>

#data
<!doctype HTml><title><!--&amp;--></title>
#errors
#document
| <!DOCTYPE HTml>
| <html>
|   <head>
|     <title>
|       "<!--&amp;-->"
|   <body>

#data
<!doctype>
#errors
No space after "doctype"
Unexpected ">"
Incorrect doctype
#document
| <!DOCTYPE >
| <html>
|   <head>
|   <body>

#data
<!---x
#errors
End of file in comment
End of file before doctype
#document
| <!-- -x -->
| <html>
|   <head>
|   <body>

#data
<body>
<div>
#errors
Unexpected start tag (body)
Expected closing tag. Unexpected end of file
#document-fragment
div
#document
| "
"
| <div>

#data
<frameset></frameset>
foo
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected non-space characters in the after frameset phase. Ignored.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
<noframes>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <frameset>
|   "
"
|   <noframes>

#data
<frameset></frameset>
<div>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected start tag (div) in the after frameset phase. Ignored.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
</html>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<frameset></frameset>
</div>
#errors
Unexpected start tag (frameset). Expected DOCTYPE.
Unexpected end tag (div) in the after frameset phase. Ignored.
#document
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<form><form>
#errors
Unexpected start tag (form). Expected DOCTYPE.
Unexpected start tag (form).
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <form>

#data
<button><button>
#errors
Unexpected start tag (button). Expected DOCTYPE.
Unexpected start tag (button) implies end tag (button).
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <button>
|     <button>

#data
<table><caption><td>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unknown error "XXX-undefined-error"
Unexpected table cell start tag (td) in the table body phase.
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|       <tbody>
|         <tr>
|           <td>

#data
<table><caption><p>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>

#data
</caption><p>
#document-fragment
caption
#errors
Unexpected end tag (caption). Ignored.
#document
| <p>

#data
<table><caption><div></caption>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (caption). Missing end tag (div).
Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <div>

#data
<table><caption></table>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end table tag in caption. Generates implied end caption.
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>

#data
<table><caption></body></col></colgroup></html></tbody></td></tfoot></th></thead></tr>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Unexpected end tag (body). Ignored.
Unexpected end tag (col). Ignored.
Unexpected end tag (colgroup). Ignored.
Unexpected end tag (html). Ignored.
Unexpected end tag (tbody). Ignored.
Unexpected end tag (td). Ignored.
Unexpected end tag (tfoot). Ignored.
Unexpected end tag (th). Ignored.
Unexpected end tag (thead). Ignored.
Unexpected end tag (tr). Ignored.
Expected closing tag. Unexpected end of file.

#data
<table><caption><p></p>
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>
#errors
Unexpected start tag (table). Expected DOCTYPE.
Expected closing tag. Unexpected end of file.
